+++ example/lib/demo/views/views.dart
@@ -1,0 +1,5 @@
+export 'default_view.dart';
+export 'favorite_and_filtered_view.dart';
+export 'favorite_view.dart';
+export 'filtered_view.dart';
+export 'demo_page.dart';