+++ lib/src/models/codes_bank.dart
@@ -2,7 +2,7 @@
 /// Contains all of the country name, code, and dial_code.
 const List<Map<String, String>> codes = [
   {
-    "name": "افغانستان",
+    "name": "Afghanistan",
     "code": "AF",
     "dial_code": "+93",
   },
@@ -17,7 +17,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+355",
   },
   {
-    "name": "الجزائر",
+    "name": "Algeria",
     "code": "DZ",
     "dial_code": "+213",
   },
@@ -57,7 +57,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+54",
   },
   {
-    "name": "Հայաստան",
+    "name": "Armenia",
     "code": "AM",
     "dial_code": "+374",
   },
@@ -72,7 +72,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+61",
   },
   {
-    "name": "Österreich",
+    "name": "Austria",
     "code": "AT",
     "dial_code": "+43",
   },
@@ -87,7 +87,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+1242",
   },
   {
-    "name": "‏البحرين",
+    "name": "Bahrain",
     "code": "BH",
     "dial_code": "+973",
   },
@@ -102,12 +102,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+1246",
   },
   {
-    "name": "Белару́сь",
+    "name": "Belarus",
     "code": "BY",
     "dial_code": "+375",
   },
   {
-    "name": "België",
+    "name": "Belgium",
     "code": "BE",
     "dial_code": "+32",
   },
@@ -127,7 +127,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+1441",
   },
   {
-    "name": "ʼbrug-yul",
+    "name": "Bhutan",
     "code": "BT",
     "dial_code": "+975",
   },
@@ -137,7 +137,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+591",
   },
   {
-    "name": "Bosna i Hercegovina",
+    "name": "Bosnia and Herzegovina",
     "code": "BA",
     "dial_code": "+387",
   },
@@ -167,7 +167,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+673",
   },
   {
-    "name": "България",
+    "name": "Bulgaria",
     "code": "BG",
     "dial_code": "+359",
   },
@@ -207,12 +207,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+1345",
   },
   {
-    "name": "Ködörösêse tî Bêafrîka",
+    "name": "Central African Republic ",
     "code": "CF",
     "dial_code": "+236",
   },
   {
-    "name": "Tchad",
+    "name": "Chad",
     "code": "TD",
     "dial_code": "+235",
   },
@@ -222,7 +222,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+56",
   },
   {
-    "name": "中国",
+    "name": "China",
     "code": "CN",
     "dial_code": "+86",
   },
@@ -247,12 +247,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+269",
   },
   {
-    "name": "République du Congo",
+    "name": "Congo",
     "code": "CG",
     "dial_code": "+242",
   },
   {
-    "name": "République démocratique du Congo",
+    "name": "Congo (Democratic Republic)",
     "code": "CD",
     "dial_code": "+243",
   },
@@ -272,7 +272,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+225",
   },
   {
-    "name": "Hrvatska",
+    "name": "Croatia",
     "code": "HR",
     "dial_code": "+385",
   },
@@ -282,17 +282,17 @@ const List<Map<String, String>> codes = [
     "dial_code": "+53",
   },
   {
-    "name": "Κύπρος",
+    "name": "Cyprus",
     "code": "CY",
     "dial_code": "+357",
   },
   {
-    "name": "Česká republika",
+    "name": "Czechia",
     "code": "CZ",
     "dial_code": "+420",
   },
   {
-    "name": "Danmark",
+    "name": "Denmark",
     "code": "DK",
     "dial_code": "+45",
   },
@@ -307,7 +307,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+1767",
   },
   {
-    "name": "República Dominicana",
+    "name": "Dominican Republic",
     "code": "DO",
     "dial_code": "+1",
   },
@@ -317,7 +317,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+593",
   },
   {
-    "name": "مصر‎",
+    "name": "Egypt",
     "code": "EG",
     "dial_code": "+20",
   },
@@ -327,32 +327,32 @@ const List<Map<String, String>> codes = [
     "dial_code": "+503",
   },
   {
-    "name": "Guinea Ecuatorial",
+    "name": "Equatorial Guinea",
     "code": "GQ",
     "dial_code": "+240",
   },
   {
-    "name": "ኤርትራ",
+    "name": "Eritrea",
     "code": "ER",
     "dial_code": "+291",
   },
   {
-    "name": "Eesti",
+    "name": "Estonia",
     "code": "EE",
     "dial_code": "+372",
   },
   {
-    "name": "ኢትዮጵያ",
+    "name": "Ethiopia",
     "code": "ET",
     "dial_code": "+251",
   },
   {
-    "name": "Falkland Islands",
+    "name": "Falkland Islands (Malvinas]",
     "code": "FK",
     "dial_code": "+500",
   },
   {
-    "name": "Føroyar",
+    "name": "Faroe Islands",
     "code": "FO",
     "dial_code": "+298",
   },
@@ -362,7 +362,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+679",
   },
   {
-    "name": "Suomi",
+    "name": "Finland",
     "code": "FI",
     "dial_code": "+358",
   },
@@ -372,17 +372,17 @@ const List<Map<String, String>> codes = [
     "dial_code": "+33",
   },
   {
-    "name": "Guyane française",
+    "name": "French Guiana",
     "code": "GF",
     "dial_code": "+594",
   },
   {
-    "name": "Polynésie française",
+    "name": "French Polynesia",
     "code": "PF",
     "dial_code": "+689",
   },
   {
-    "name": "Territoire des Terres australes et antarctiques fr",
+    "name": "French Southern Territories",
     "code": "TF",
     "dial_code": "+262",
   },
@@ -397,12 +397,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+220",
   },
   {
-    "name": "საქართველო",
+    "name": "Georgia",
     "code": "GE",
     "dial_code": "+995",
   },
   {
-    "name": "Deutschland",
+    "name": "Germany",
     "code": "DE",
     "dial_code": "+49",
   },
@@ -417,12 +417,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+350",
   },
   {
-    "name": "Ελλάδα",
+    "name": "Greece",
     "code": "GR",
     "dial_code": "+30",
   },
   {
-    "name": "Kalaallit Nunaat",
+    "name": "Greenland",
     "code": "GL",
     "dial_code": "+299",
   },
@@ -452,12 +452,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+44",
   },
   {
-    "name": "Guinée",
+    "name": "Guinea",
     "code": "GN",
     "dial_code": "+224",
   },
   {
-    "name": "Guiné-Bissau",
+    "name": "Guinea-Bissau",
     "code": "GW",
     "dial_code": "+245",
   },
@@ -467,7 +467,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+592",
   },
   {
-    "name": "Haïti",
+    "name": "Haiti",
     "code": "HT",
     "dial_code": "+509",
   },
@@ -477,7 +477,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+0",
   },
   {
-    "name": "Vaticano",
+    "name": "Vatican",
     "code": "VA",
     "dial_code": "+379",
   },
@@ -487,22 +487,22 @@ const List<Map<String, String>> codes = [
     "dial_code": "+504",
   },
   {
-    "name": "香港",
+    "name": "Hong Kong",
     "code": "HK",
     "dial_code": "+852",
   },
   {
-    "name": "Magyarország",
+    "name": "Hungary",
     "code": "HU",
     "dial_code": "+36",
   },
   {
-    "name": "Ísland",
+    "name": "Iceland",
     "code": "IS",
     "dial_code": "+354",
   },
   {
-    "name": "भारत",
+    "name": "India",
     "code": "IN",
     "dial_code": "+91",
   },
@@ -512,17 +512,17 @@ const List<Map<String, String>> codes = [
     "dial_code": "+62",
   },
   {
-    "name": "ایران",
+    "name": "Iran",
     "code": "IR",
     "dial_code": "+98",
   },
   {
-    "name": "العراق",
+    "name": "Iraq",
     "code": "IQ",
     "dial_code": "+964",
   },
   {
-    "name": "Éire",
+    "name": "Ireland",
     "code": "IE",
     "dial_code": "+353",
   },
@@ -532,12 +532,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+44",
   },
   {
-    "name": "ישראל",
+    "name": "Israel",
     "code": "IL",
     "dial_code": "+972",
   },
   {
-    "name": "Italia",
+    "name": "Italy",
     "code": "IT",
     "dial_code": "+39",
   },
@@ -547,7 +547,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+1876",
   },
   {
-    "name": "日本",
+    "name": "Japan",
     "code": "JP",
     "dial_code": "+81",
   },
@@ -557,12 +557,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+44",
   },
   {
-    "name": "الأردن",
+    "name": "Jordan",
     "code": "JO",
     "dial_code": "+962",
   },
   {
-    "name": "Қазақстан",
+    "name": "Kazakhstan",
     "code": "KZ",
     "dial_code": "+7",
   },
@@ -577,42 +577,37 @@ const List<Map<String, String>> codes = [
     "dial_code": "+686",
   },
   {
-    "name": "북한",
+    "name": "Korea Democratic People's Republic",
     "code": "KP",
     "dial_code": "+850",
   },
   {
-    "name": "대한민국",
+    "name": "Korea",
     "code": "KR",
     "dial_code": "+82",
   },
   {
-    "name": "Republika e Kosovës",
-    "code": "XK",
-    "dial_code": "+383",
-  },
-  {
-    "name": "الكويت",
+    "name": "Kuwait",
     "code": "KW",
     "dial_code": "+965",
   },
   {
-    "name": "Кыргызстан",
+    "name": "Kyrgyzstan",
     "code": "KG",
     "dial_code": "+996",
   },
   {
-    "name": "ສປປລາວ",
+    "name": "Lao People's Democratic Republic",
     "code": "LA",
     "dial_code": "+856",
   },
   {
-    "name": "Latvija",
+    "name": "Latvia",
     "code": "LV",
     "dial_code": "+371",
   },
   {
-    "name": "لبنان",
+    "name": "Lebanon",
     "code": "LB",
     "dial_code": "+961",
   },
@@ -627,7 +622,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+231",
   },
   {
-    "name": "‏ليبيا",
+    "name": "Libya",
     "code": "LY",
     "dial_code": "+218",
   },
@@ -637,7 +632,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+423",
   },
   {
-    "name": "Lietuva",
+    "name": "Lithuania",
     "code": "LT",
     "dial_code": "+370",
   },
@@ -647,17 +642,17 @@ const List<Map<String, String>> codes = [
     "dial_code": "+352",
   },
   {
-    "name": "澳門",
+    "name": "Macao",
     "code": "MO",
     "dial_code": "+853",
   },
   {
-    "name": "Македонија",
+    "name": " Macedonia (North Republic)",
     "code": "MK",
     "dial_code": "+389",
   },
   {
-    "name": "Madagasikara",
+    "name": "Madagascar",
     "code": "MG",
     "dial_code": "+261",
   },
@@ -687,7 +682,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+356",
   },
   {
-    "name": "M̧ajeļ",
+    "name": "Marshall Islands",
     "code": "MH",
     "dial_code": "+692",
   },
@@ -697,12 +692,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+596",
   },
   {
-    "name": "موريتانيا",
+    "name": "Mauritania",
     "code": "MR",
     "dial_code": "+222",
   },
   {
-    "name": "Maurice",
+    "name": "Mauritius",
     "code": "MU",
     "dial_code": "+230",
   },
@@ -732,12 +727,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+377",
   },
   {
-    "name": "Монгол улс",
+    "name": "Mongolia",
     "code": "MN",
     "dial_code": "+976",
   },
   {
-    "name": "Црна Гора",
+    "name": "Montenegro",
     "code": "ME",
     "dial_code": "+382",
   },
@@ -747,12 +742,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+1664",
   },
   {
-    "name": "المغرب",
+    "name": "Morocco",
     "code": "MA",
     "dial_code": "+212",
   },
   {
-    "name": "Moçambique",
+    "name": "Mozambique",
     "code": "MZ",
     "dial_code": "+258",
   },
@@ -772,12 +767,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+674",
   },
   {
-    "name": "नेपाल",
+    "name": "Nepal",
     "code": "NP",
     "dial_code": "+977",
   },
   {
-    "name": "Nederland",
+    "name": "Netherlands",
     "code": "NL",
     "dial_code": "+31",
   },
@@ -787,7 +782,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+599",
   },
   {
-    "name": "Nouvelle-Calédonie",
+    "name": "New Caledonia",
     "code": "NC",
     "dial_code": "+687",
   },
@@ -812,7 +807,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+234",
   },
   {
-    "name": "Niuē",
+    "name": "Niue",
     "code": "NU",
     "dial_code": "+683",
   },
@@ -827,12 +822,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+1670",
   },
   {
-    "name": "Norge",
+    "name": "Norway",
     "code": "NO",
     "dial_code": "+47",
   },
   {
-    "name": "عمان",
+    "name": "Oman",
     "code": "OM",
     "dial_code": "+968",
   },
@@ -847,7 +842,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+680",
   },
   {
-    "name": "فلسطين",
+    "name": "Palestine",
     "code": "PS",
     "dial_code": "+970",
   },
@@ -882,7 +877,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+64",
   },
   {
-    "name": "Polska",
+    "name": "Poland",
     "code": "PL",
     "dial_code": "+48",
   },
@@ -902,17 +897,17 @@ const List<Map<String, String>> codes = [
     "dial_code": "+1787",
   },
   {
-    "name": "قطر",
+    "name": "Qatar",
     "code": "QA",
     "dial_code": "+974",
   },
   {
-    "name": "România",
+    "name": "Romania",
     "code": "RO",
     "dial_code": "+40",
   },
   {
-    "name": "Россия",
+    "name": "Russian Federation",
     "code": "RU",
     "dial_code": "+7",
   },
@@ -952,7 +947,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+590",
   },
   {
-    "name": "Saint-Pierre-et-Miquelon",
+    "name": "Saint Pierre and Miquelon",
     "code": "PM",
     "dial_code": "+508",
   },
@@ -972,22 +967,22 @@ const List<Map<String, String>> codes = [
     "dial_code": "+378",
   },
   {
-    "name": "São Tomé e Príncipe",
+    "name": "Sao Tome and Principe",
     "code": "ST",
     "dial_code": "+239",
   },
   {
-    "name": "العربية السعودية",
+    "name": "Saudi Arabia",
     "code": "SA",
     "dial_code": "+966",
   },
   {
-    "name": "Sénégal",
+    "name": "Senegal",
     "code": "SN",
     "dial_code": "+221",
   },
   {
-    "name": "Србија",
+    "name": "Serbia",
     "code": "RS",
     "dial_code": "+381",
   },
@@ -1007,12 +1002,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+65",
   },
   {
-    "name": "Slovensko",
+    "name": "Slovakia",
     "code": "SK",
     "dial_code": "+421",
   },
   {
-    "name": "Slovenija",
+    "name": "Slovenia",
     "code": "SI",
     "dial_code": "+386",
   },
@@ -1022,7 +1017,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+677",
   },
   {
-    "name": "Soomaaliya",
+    "name": "Somalia",
     "code": "SO",
     "dial_code": "+252",
   },
@@ -1042,7 +1037,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+500",
   },
   {
-    "name": "España",
+    "name": "Spain",
     "code": "ES",
     "dial_code": "+34",
   },
@@ -1052,7 +1047,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+94",
   },
   {
-    "name": "السودان",
+    "name": "Sudan",
     "code": "SD",
     "dial_code": "+249",
   },
@@ -1062,37 +1057,37 @@ const List<Map<String, String>> codes = [
     "dial_code": "+597",
   },
   {
-    "name": "Svalbard og Jan Mayen",
+    "name": "Svalbard and Jan Mayen",
     "code": "SJ",
     "dial_code": "+47",
   },
   {
-    "name": "Swaziland",
+    "name": "Eswatini",
     "code": "SZ",
     "dial_code": "+268",
   },
   {
-    "name": "Sverige",
+    "name": "Sweden",
     "code": "SE",
     "dial_code": "+46",
   },
   {
-    "name": "Schweiz",
+    "name": "Switzerland",
     "code": "CH",
     "dial_code": "+41",
   },
   {
-    "name": "سوريا",
+    "name": "Syrian Arab Republic",
     "code": "SY",
     "dial_code": "+963",
   },
   {
-    "name": "臺灣",
+    "name": "Taiwan",
     "code": "TW",
     "dial_code": "+886",
   },
   {
-    "name": "Тоҷикистон",
+    "name": "Tajikistan",
     "code": "TJ",
     "dial_code": "+992",
   },
@@ -1102,7 +1097,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+255",
   },
   {
-    "name": "ประเทศไทย",
+    "name": "Thailand",
     "code": "TH",
     "dial_code": "+66",
   },
@@ -1132,17 +1127,17 @@ const List<Map<String, String>> codes = [
     "dial_code": "+1868",
   },
   {
-    "name": "تونس",
+    "name": "Tunisia",
     "code": "TN",
     "dial_code": "+216",
   },
   {
-    "name": "Türkiye",
+    "name": "Turkey",
     "code": "TR",
     "dial_code": "+90",
   },
   {
-    "name": "Türkmenistan",
+    "name": "Turkmenistan",
     "code": "TM",
     "dial_code": "+993",
   },
@@ -1162,12 +1157,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+256",
   },
   {
-    "name": "Україна",
+    "name": "Ukraine",
     "code": "UA",
     "dial_code": "+380",
   },
   {
-    "name": "دولة الإمارات العربية المتحدة",
+    "name": "United Arab Emirates",
     "code": "AE",
     "dial_code": "+971",
   },
@@ -1187,7 +1182,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+598",
   },
   {
-    "name": "O‘zbekiston",
+    "name": "Uzbekistan",
     "code": "UZ",
     "dial_code": "+998",
   },
@@ -1202,7 +1197,7 @@ const List<Map<String, String>> codes = [
     "dial_code": "+58",
   },
   {
-    "name": "Việt Nam",
+    "name": "Viet Nam",
     "code": "VN",
     "dial_code": "+84",
   },
@@ -1217,12 +1212,12 @@ const List<Map<String, String>> codes = [
     "dial_code": "+1340",
   },
   {
-    "name": "Wallis et Futuna",
+    "name": "Wallis and Futuna",
     "code": "WF",
     "dial_code": "+681",
   },
   {
-    "name": "اليَمَن",
+    "name": "Yemen",
     "code": "YE",
     "dial_code": "+967",
   },