+++ lib/src/widgets/country_code_picker_modal.dart
@@ -10,24 +10,59 @@ import 'package:flutter/material.dart';
 /// {@endtemplate}
 class CountryCodePickerModal extends StatefulWidget {
   /// {@macro country_code_picker_modal}
-  const CountryCodePickerModal({Key? key}) : super(key: key);
+  const CountryCodePickerModal({
+    Key? key,
+    this.favorites = const [],
+    this.filteredCountries = const [],
+    required this.customIcon,
+    required this.showSearchBar,
+  }) : super(key: key);
+
+  /// {@macro favorites}
+  final List<String> favorites;
+
+  /// {@macro filtered_countries}
+  final List<String> filteredCountries;
+
+  /// {@macro favorite_icon}
+  final Icon customIcon;
+
+  /// {@macro show_search_bar}
+  final bool showSearchBar;
 
   @override
   State<CountryCodePickerModal> createState() => _CountryCodePickerModalState();
 }
 
 class _CountryCodePickerModalState extends State<CountryCodePickerModal> {
-  final allCountryCodes = <CountryCode>[];
+  late final List<CountryCode> baseList;
   final availableCountryCodes = <CountryCode>[];
+
   late TextEditingController controller;
   @override
   void initState() {
-    controller = TextEditingController();
-    for (final code in codes) {
-      allCountryCodes.add(CountryCode.fromMap(code));
-    }
-    availableCountryCodes.addAll(allCountryCodes);
+    _initCountries();
     super.initState();
+  }
+
+  void _initCountries() {
+    final allCountryCodes = codes.map(CountryCode.fromMap).toList();
+    controller = TextEditingController();
+
+    final favoriteList = <CountryCode>[
+      if (widget.favorites.isNotEmpty)
+        ...allCountryCodes.where((c) => widget.favorites.contains(c.code))
+    ];
+    final filteredList = [
+      ...widget.filteredCountries.isNotEmpty
+          ? allCountryCodes.where(
+              (c) => widget.filteredCountries.contains(c.code),
+            )
+          : allCountryCodes,
+    ]..removeWhere((c) => widget.favorites.contains(c.code));
+
+    baseList = [...favoriteList, ...filteredList];
+    availableCountryCodes.addAll(baseList);
   }
 
   @override
@@ -43,44 +78,47 @@ class _CountryCodePickerModalState extends State<CountryCodePickerModal> {
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
         const _ModalTitle(),
-        Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16),
-          child: TextField(
-            keyboardType: TextInputType.text,
-            decoration: const InputDecoration(
-              hintText: "'Country', 'Code' or 'Dial Code'",
-              hintStyle: TextStyle(fontSize: 12),
-              suffixIcon: Icon(Icons.search),
-              border: OutlineInputBorder(
-                borderRadius: BorderRadius.all(borderRadius),
-                borderSide: BorderSide(
-                  width: 2,
-                  style: BorderStyle.none,
-                ),
-              ),
-              filled: true,
-              contentPadding: EdgeInsets.symmetric(horizontal: 16),
-              fillColor: Colors.white,
-            ),
-            onChanged: (query) {
-              availableCountryCodes
-                ..clear()
-                ..addAll(
-                  List<CountryCode>.from(
-                    allCountryCodes.where(
-                      (c) =>
-                          c.code.toLowerCase().contains(query.toLowerCase()) ||
-                          c.dialCode
-                              .toLowerCase()
-                              .contains(query.toLowerCase()) ||
-                          c.name.toLowerCase().contains(query.toLowerCase()),
-                    ),
+        if (widget.showSearchBar)
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 16),
+            child: TextField(
+              keyboardType: TextInputType.text,
+              decoration: const InputDecoration(
+                hintText: "'Country', 'Code' or 'Dial Code'",
+                hintStyle: TextStyle(fontSize: 12),
+                suffixIcon: Icon(Icons.search),
+                border: OutlineInputBorder(
+                  borderRadius: BorderRadius.all(borderRadius),
+                  borderSide: BorderSide(
+                    width: 2,
+                    style: BorderStyle.none,
                   ),
-                );
-              setState(() {});
-            },
+                ),
+                filled: true,
+                contentPadding: EdgeInsets.symmetric(horizontal: 16),
+                fillColor: Colors.white,
+              ),
+              onChanged: (query) {
+                availableCountryCodes
+                  ..clear()
+                  ..addAll(
+                    List<CountryCode>.from(
+                      baseList.where(
+                        (c) =>
+                            c.code
+                                .toLowerCase()
+                                .contains(query.toLowerCase()) ||
+                            c.dialCode
+                                .toLowerCase()
+                                .contains(query.toLowerCase()) ||
+                            c.name.toLowerCase().contains(query.toLowerCase()),
+                      ),
+                    ),
+                  );
+                setState(() {});
+              },
+            ),
           ),
-        ),
         Expanded(
           child: ListView(
             children: [
@@ -89,12 +127,46 @@ class _CountryCodePickerModalState extends State<CountryCodePickerModal> {
                   onTap: () => Navigator.pop(context, code),
                   leading: code.flagImage,
                   title: Text(code.name),
-                  trailing: Text(code.dialCode),
+                  trailing: _ListTrailing(
+                    code: code,
+                    favorites: widget.favorites,
+                    icon: widget.customIcon,
+                  ),
                 )
             ],
           ),
         ),
       ],
+    );
+  }
+}
+
+class _ListTrailing extends StatelessWidget {
+  const _ListTrailing({
+    Key? key,
+    required this.code,
+    required this.favorites,
+    required this.icon,
+  }) : super(key: key);
+  final CountryCode code;
+  final List<String> favorites;
+  final Icon icon;
+
+  @override
+  Widget build(BuildContext context) {
+    if (favorites.isEmpty) return Text(code.dialCode);
+
+    final index = favorites.indexWhere((f) => f == code.code);
+    final iconWidth = MediaQuery.of(context).size.width * 0.2;
+    return SizedBox(
+      width: iconWidth,
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          Text(code.dialCode),
+          if (index != -1) icon,
+        ],
+      ),
     );
   }
 }