+++ test/fl_country_code_picker_test.dart
@@ -1,18 +1,12 @@
 // ignore_for_file: prefer_const_constructors
 import 'package:fl_country_code_picker/fl_country_code_picker.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mocktail/mocktail.dart';
-
-class MockBuildContext extends Mock implements BuildContext {}
 
 void main() {
   late FlCountryCodePicker picker;
-  late MockBuildContext _mockContext;
 
   setUp(() {
-    _mockContext = MockBuildContext();
-    picker = FlCountryCodePicker(context: _mockContext);
+    picker = FlCountryCodePicker();
   });
 
   group('FlCountryCodePicker', () {
@@ -23,7 +17,7 @@ void main() {
     group('countryCodes', () {
       test('gets list of country codes', () {
         expect(
-          picker.countryCodes().runtimeType,
+          picker.countryCodes.runtimeType,
           List<CountryCode>,
         );
       });