+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
 
 Initial release
+
+## 0.0.2
+
+- refactor: make class constable
+- refactor: enable access to `CountryCodes` list without using context
+- feat: add option for hiding the search bar
+- feat: add support for favorite countries ([#1](https://github.com/fernan542/fl_country_code_picker/issues/1))
+- feat: add support for countries filtering ([#1](https://github.com/fernan542/fl_country_code_picker/issues/1))